data_collection/requirements.txt
alembic
annotated-types
anyio
bcrypt
certifi
charset-normalizer
click
dnspython
email-validator
exceptiongroup
fastapi
greenlet
h11
idna
Jinja2
Mako
MarkupSafe
psycopg2-binary
pydantic
pydantic_core
PyJWT
python-dotenv
python-multipart
requests
redis
six
sniffio
SQLAlchemy
starlette
typing_extensions
urllib3
uvicorn
pytz
fastapi-utils